refactor(calculator): CalculatorBasicInputGroup으로 컴포넌트 쪼개기
