+++ app/src/Calculator/index.tsx
@@ -1,5 +1,5 @@
 import { useQuery } from '@apollo/client';
-import { useAtom } from 'jotai';
+import { useSetAtom } from 'jotai';
 import { useEffect } from 'react';
 
 import { calculatorPropsAtom } from '@/Calculator/atoms/calculatorPropsAtom';
@@ -12,22 +12,14 @@ import {
 import { Footer } from '@core/components/Footer';
 import { gql } from '@shared/__generated__';
 import { DashboardTemp } from '@shared/components/Dashboard/DashboardTemp';
-import { InfoTooltip } from '@shared/components/InfoTooltip';
 import { Seo } from '@shared/components/Seo';
-import {
-  Body1MediumText,
-  H1BoldText,
-  HStack,
-  Input,
-  Text,
-  VStack,
-} from '@shared/ui-kit';
+import { H1BoldText, VStack } from '@shared/ui-kit';
 import { getBlackholeDaysLeft } from '@shared/utils/getBlackholeDaysLeft';
 import { getTimeDiffFromNow } from '@shared/utils/getTimeDiffFromNow';
 import { useDeviceType } from '@shared/utils/react-responsive/useDeviceType';
 
 import { subjectListAtom } from './atoms/subjectListAtom';
-import { useSubjectList } from './hooks/useSubjectList';
+import { CalculatorBasicInfoInputGroup } from './components/CalculatorBasicInfoInputGroup';
 
 export const GET_BLACKHOLE_INFO = gql(/* GraphQL */ `
   query GetBlackholeInfo {
@@ -46,28 +38,10 @@ export const GET_BLACKHOLE_INFO = gql(/* GraphQL */ `
 `);
 
 const CalculatorPage = () => {
-  const [calculatorProps, setCalculatorProps] = useAtom(calculatorPropsAtom);
-  const [subjectList, setSubjectList] = useAtom(subjectListAtom);
-  const { updateSubjectList } = useSubjectList();
-  const { currentLevel, daysFromStart, currentBlackhole } = calculatorProps;
+  const setCalculatorProps = useSetAtom(calculatorPropsAtom);
+  const setSubjectList = useSetAtom(subjectListAtom);
   const device = useDeviceType();
   const { loading, error, data } = useQuery(GET_BLACKHOLE_INFO);
-
-  const handleChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = Number(e.target.value);
-    if (isNaN(value) || value < 0) return;
-    const name = e.target.name as keyof typeof calculatorProps;
-    setCalculatorProps((prev) => ({
-      ...prev,
-      [name]: value,
-    }));
-  };
-
-  //handleChange에서 updateSubjectList를 호출할 때 update되지 않는 문제
-  useEffect(() => {
-    updateSubjectList(subjectList);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [calculatorProps]);
 
   useEffect(() => {
     if (!data) {
@@ -118,56 +92,7 @@ const CalculatorPage = () => {
       <VStack w="100%" spacing="4rem">
         <VStack w="100%" align="start" spacing="1rem">
           <H1BoldText>블랙홀 계산기</H1BoldText>
-          <HStack spacing="2rem">
-            <HStack w="13rem" justify="start" spacing="1rem">
-              <Body1MediumText>현재 레벨</Body1MediumText>
-              <InfoTooltip text="레벨이 8.41을 넘으면, 블랙홀 기간이 늘지 않아요." />
-            </HStack>
-            <Input
-              name="currentLevel"
-              type="number"
-              min="0"
-              max="30"
-              step="0.01"
-              value={currentLevel}
-              onChange={handleChange}
-              style={{ width: '5rem' }}
-            />
-          </HStack>
-          <HStack spacing="2rem">
-            <HStack w="13rem" justify="start" spacing="1rem">
-              <Body1MediumText>현재 블랙홀</Body1MediumText>
-              <InfoTooltip text="현재 블랙홀 + 본 과정 시작 날짜가 670일이 넘으면, 블랙홀 기간이 늘지 않아요." />
-            </HStack>
-            <HStack spacing="0.3rem">
-              <Input
-                name="currentBlackhole"
-                type="number"
-                min="0"
-                value={currentBlackhole}
-                onChange={handleChange}
-                style={{ width: '5rem' }}
-              />
-              <Text>일</Text>
-            </HStack>
-          </HStack>
-          <HStack spacing="2rem">
-            <HStack w="13rem" justify="start" spacing="1rem">
-              <Body1MediumText>본과정 시작한지</Body1MediumText>
-              <InfoTooltip text="휴학일이 포함된 경우, 휴학 기간을 뺄 수 있어요." />
-            </HStack>
-            <HStack spacing="0.3rem">
-              <Input
-                name="daysFromStart"
-                type="number"
-                min="0"
-                value={daysFromStart}
-                onChange={handleChange}
-                style={{ width: '5rem' }}
-              />
-              <Text>일</Text>
-            </HStack>
-          </HStack>
+          <CalculatorBasicInfoInputGroup />
         </VStack>
         <DashboardTemp
           contents={calculatorPageDashboardContents}
